requirements.txt
progressbar2
requests
ijson